requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
djangorestframework==3.14.0
Faker==15.3.3
greenlet==2.0.1
numpy==1.23.5
pandas==1.5.2
pkg_resources==0.0.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
